working on building a provider.js class that returns a promise
