+++ server/db/models/practice.spec.js
@@ -1,0 +1,10 @@
+const { Practice } = require('./');
+
+const practice = new Practice();
+
+const readPractices = async () => {
+  const data = await practice.findAll();
+  console.log(data);
+};
+
+readPractices();
