--- server/db/models/provider.spec.js
@@ -1,5 +1,0 @@
-const { Provider } = require('./');
-
-const provider = new Provider();
-
-console.log(provider.findAll());
